Hold up, those weren't actually bugs.
I'm sorry. When I "fixed" the theme, I didn't realize that the theme was heavily reliant on the rc.lua included, which I didn't use, and I kept the rc.lua here, which is probably going to break the relationship. Here is my rc.lua which I use with this theme, this is basically my personal config at this point. It adds the corner nw, ne, sw, se layouts, and moves the panel to the bottom of the screen.