rc.lua 2.0 as well as my own screenshot.
I made some changes to my own rc.lua for awesomeWM 4.3-2 or 4.x:
    Added keybinds for keys I use a lot.
    Added screenshot functionality that wasn't there before.
This is not a fork of Igaggini's rc.lua, only the theme is. I modified the theme to be compatible with my rc.lua.
I also added my own screenshot and replaced it in the README.md.