rc.lua patch adding slock
I am really sorry that I added more commits for this fork than I really needed to. This is a patch for the rc.lua configuration file that adds the ability to lock the screen using Mod+Esc using slock, and I noticed some old comments containing the string "removeiferror" I used in order to quickly jump to places that were likely to have bugs, a trick I used for debugging, so I removed those in this version because they're irrelevant.